+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- repeatable set form elements to form element mappings
+
 ## [13.1.3] - 2026-04-23
 
 ### Dependencies